# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PM Network API - GitHub for Product Managers
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/decks:
    get:
      operationId: getUserDecks
      tags: [users]
      summary: Get decks by user
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: User decks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deck"
  /users/{id}/starred:
    get:
      operationId: getUserStarred
      tags: [users]
      summary: Get decks starred by user
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Starred decks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deck"
  /users/{id}/follow:
    post:
      operationId: followUser
      tags: [users]
      summary: Follow a user
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FollowInput"
      responses:
        "200":
          description: Follow toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"

  # ─── Decks ────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, name, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        name: { type: string }
        bio: { type: ["string", "null"] }
        title: { type: ["string", "null"] }
        company: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        followers: { type: integer }
        following: { type: integer }
        deckCount: { type: integer }
        createdAt: { type: string, format: date-time }
    UserInput:
      type: object
      required: [username, name]
      properties:
        username: { type: string, minLength: 2 }
        name: { type: string, minLength: 1 }
        bio: { type: string }
        title: { type: string }
        company: { type: string }
        avatarUrl: { type: string }
    UserUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        bio: { type: string }
        title: { type: string }
        company: { type: string }
        avatarUrl: { type: string }
    FollowInput:
      type: object
      required: [followerId]
      properties:
        followerId: { type: integer }
    FollowResult:
      type: object
      required: [following]
      properties:
        following: { type: boolean }
        followerCount: { type: integer }

    # ─── Deck ──────────────────────────────────────────────
    Deck:
      type: object
      required: [id, userId, title, stars, views, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        authorName: { type: string }
        authorUsername: { type: string }
        authorAvatarUrl: { type: ["string", "null"] }
        title: { type: string }
        description: { type: ["string", "null"] }
        thumbnailUrl: { type: ["string", "null"] }
        fileUrl: { type: ["string", "null"] }
        tags: { type: array, items: { type: string } }
        stars: { type: integer }
        views: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...