# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PM Network API - GitHub for Product Managers
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  name: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  followers?: number;
  following?: number;
  deckCount?: number;
  createdAt: Date;
}