# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PM Network API - GitHub for Product Managers
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  /** @minLength 1 */
  name?: string;
  bio?: string;
  title?: string;
  company?: string;
  avatarUrl?: string;
}